+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email 'ignasi.barrera@abiquo.com'
 license          'Apache 2.0'
 description      'Installs and configures an Abiquo platform'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.0'
+version          '0.2.0'
 
 recipe 'abiquo', 'Installs and configures an Abiquo platform'
 recipe 'abiquo::upgrade', 'Upgrades an Abiquo platform'