+++ recipes/install_remoteservices.rb
@@ -30,13 +30,6 @@ include_recipe "redisio::enable"
     end
 end
 
-# The abiquo-release-ee package installs this repo. As we are in control
-# of the created repos, we just delete it, to avoid having it conflict with
-# the configured ones.
-yum_repository "Abiquo-Base" do
-    action :delete
-end
-
 selinux_state "SELinux Permissive" do
     action :permissive
 end
@@ -46,8 +39,4 @@ iptables_rule "firewall-tomcat"
 
 service "rpcbind" do
     action [:enable, :start]
-end
-
-service "abiquo-tomcat" do
-    start_command "service abiquo-tomcat jpda" if node['abiquo']['tomcat-jpda']
 end