+++ recipes/setup_kvm.rb
@@ -39,7 +39,5 @@ template "/etc/abiquo-aim.ini" do
 end
 
 service "abiquo-aim" do
-    # The abiquo-aim script hardly ever returns 0 :(
-    ignore_failure true
     action [:enable, :start]
 end