+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 abiquo CHANGELOG
 ================
 
+## 0.2.0
+
+* Support Java 8.
+* Support using JCE unlimited strength encryption policies.
+* Allow customization of the RabbitMQ and Redis properties.
+* Dropped support to configure the Abiquo Tomcat in jpda mode.
+
 ## 0.1.0
 
 * Initial release. Support for Monolithic installations from scratch and platform upgrades using nightly builds.